+++ src/host/layer23/src/common/networks.c
@@ -1787,7 +1787,7 @@ int gsm_match_mcc(uint16_t mcc, char *imsi)
 }
 
 /* GSM 03.22 Annex A */
-int gsm_match_mnc(uint16_t mcc, uint8_t mnc, char *imsi)
+int gsm_match_mnc(uint16_t mcc, uint16_t mnc, char *imsi)
 {
 	uint16_t sim_mnc;
